requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
Flask==3.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
numpy==1.26.4
packaging==24.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.23
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
Werkzeug==3.0.1
WTForms==3.2.1
